refactor: infrastructure 패키지 정리
